requirements.txt
streamlit
pdfplumber
python-docx
PyPDF2
requests
openai 